requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
numpy
streamlit
streamlit-chat
sentencepiece
gradio
auto_gptq
onnxruntime
nncf>=2.8.1
optimum-intel>=1.15.2
transformers>=4.37.2
# openvino-nightly
openvino>=2023.3.0
onnx>=1.15.0
huggingface-hub>=0.21.3